finished jobs handlers

+++ routes/jobs.js
@@ -9,10 +9,30 @@ const {
 } = require('../middleware/auth');
 const { validate } = require('jsonschema');
 const jobSchema = require('../jsonSchema/jobs');
+const APIError = require('../APIError');
 
 router.get('', ensureloggedin, async function(req, res, next) {
   try {
-    const data = await db.query('SELECT * FROM jobs');
+    const offset = req.query.offset ? req.query.offset : 0;
+    const limit =
+      req.query.limit && req.query.limit < 50 ? req.query.limit : 50;
+
+    const search = req.query.search ? req.query.search + '%' : req.query.search;
+    let data;
+
+    if (!search) {
+      data = await db.query('SELECT * FROM jobs LIMIT $1 OFFSET $2', [
+        limit,
+        offset
+      ]);
+    } else {
+      data = await db.query(
+        'SELECT * FROM jobs WHERE handle ILIKE $1 LIMIT $2 OFFSET $3',
+        [search, limit, offset]
+      );
+    }
+    // const data = await db.query('SELECT * FROM jobs');
+    // console.log(data);
     return res.json(data.rows);
   } catch (err) {
     return next(err);
@@ -24,14 +44,18 @@ router.post('', ensureLoginCompany, async function(req, res, next) {
     const result = validate(req.body, jobSchema);
     if (!result.valid) {
       // pass the validation errors to the error handler
-      return next(result.errors.map(e => e.stack));
+      return next(
+        new APIError(
+          400,
+          'Bad Request',
+          result.errors.map(e => e.stack).join('. ')
+        )
+      );
     }
-    const token = req.headers.authorization;
-    const decodedToken = jwt.verify(token, 'SECRET');
-    const companyid = decodedToken.company_id;
+
     const data = await db.query(
-      'INSERT INTO jobs (title, salary, equity, company_id) VALUES($1, $2, $3, $4) RETURNING *',
-      [req.body.title, req.body.salary, req.body.equity, companyid]
+      'INSERT INTO jobs (title, salary, equity, company) VALUES($1, $2, $3, $4) RETURNING *',
+      [req.body.title, req.body.salary, req.body.equity, req.handle]
     );
     return res.json(data.rows[0]);
   } catch (err) {
@@ -44,30 +68,51 @@ router.get('/:id', ensureloggedin, async function(req, res, next) {
     const data = await db.query('SELECT * FROM jobs where id=$1', [
       req.params.id
     ]);
+
     return res.json(data.rows[0]);
   } catch (err) {
     return next(err);
   }
 });
 
-router.patch('/:id', ensureCorrectCompany, async function(req, res, next) {
+// job/:id
+router.patch('/:id', ensureLoginCompany, async function(req, res, next) {
   try {
+    // SELECT * FROM jobs WHERE id = $1;
+    const job = await db.query('SELECT * FROM jobs WHERE id=$1', [
+      req.params.id
+    ]);
+    // compare req.company_handle with job.company
+
     const result = validate(req.body, jobSchema);
+    if (job.rows[0].company !== req.handle) {
+      return next(
+        new APIError(
+          401,
+          'Unauthorized',
+          result.errors.map(e => e.stack).join('. ')
+        )
+      );
+    }
+    //  if its not the same company handle, next(err)
     if (!result.valid) {
       // pass the validation errors to the error handler
-      return next(result.errors.map(e => e.stack));
+      return next(
+        new APIError(
+          400,
+          'Bad Request',
+          result.errors.map(e => e.stack).join('. ')
+        )
+      );
     }
-    const token = req.headers.authorization;
-    const decodedToken = jwt.verify(token, 'SECRET');
-    const companyid = decodedToken.company_id;
-
+    // console.log('HANDLE', req.handle);
     const data = await db.query(
-      'UPDATE jobs SET title=$1, salary=$2, equity=$3, company_id=$4 WHERE id=$5 RETURNING *',
+      'UPDATE jobs SET title=$1, salary=$2, equity=$3, company=$4 WHERE id=$5 RETURNING *',
       [
         req.body.title,
         req.body.salary,
         req.body.equity,
-        companyid,
+        req.handle,
         req.params.id
       ]
     );
@@ -77,8 +122,20 @@ router.patch('/:id', ensureCorrectCompany, async function(req, res, next) {
   }
 });
 
-router.delete('/:id', ensureCorrectCompany, async function(req, res, next) {
+router.delete('/:id', ensureLoginCompany, async function(req, res, next) {
   try {
+    const job = await db.query('SELECT * FROM jobs WHERE id=$1', [
+      req.params.id
+    ]);
+    // compare req.company_handle with job.company
+
+    //const result = validate(req.body, jobSchema);
+    if (job.rows[0].company !== req.handle) {
+      return next(
+        new APIError(401, 'Unauthorized', 'Cannot delete other jobs')
+      );
+    }
+
     const data = await db.query('DELETE FROM jobs WHERE id=$1', [
       req.params.id
     ]);
